// src/app/components/pages-sidebar/misure-cem/misure-cem-dto.ts
export class MisureCemDto {
    attività?: any;

    idimiscem?: any;
    sopralluogo?: string;
    tipologia?: string;
    comune?: string;
    provincia?: string;
    codsito?: string;
    data?: any;
    modstrum?: string;
    sonda?: string;
    tartstrum?: string;
    puntodimisura?: string;
    descpuntomis?: string;
    latitudine?: string;
    longitudine?: string;
    valeffrmsvm?: string;
    valeffrmsut?: string;
    permcontgio?: string;
    limleggevm?: string;
    limleggeut?: string;
    rangefreq?: string;
    gammamisura?: string;
    sovraccarico?: string;
    risoluzione?: string;
    misisotrop?: string;
}